==> notebooks/02_data_processing.ipynb <==
This is a record of the initial Data Processing when the project was first began. As I went through the project I decided that ranking.csv is all based on cumulative averages and is not really needed as these and other rolling averages can be calculated from the games.csv data. I also decided to hold off initially using the individual player stats in games_details.csv. I may add select features from this data later.

The functions developed in this notebook were added to data_processing.py in the src folder for reusability

# Data Processing

### Summary

Games.csv and ranking.csv will be merged after initial data processing. Since all the features are *post* game data (final score, winning percentage after the game, etc...) they cannot be used as predictors for the current game. All the features will be used as predictors for the "next game", so the data will need to adjusted so that the TARGET (HOME_TEAM_WINS) is in the same row as the predictors. 

The easiest approach seems to be add a field called TARGET that denotes whether the home team won its *next* game or not.

Game_details.csv will initially be held in reserve for feature engineering. With a roster of 24 players per game and 21 features per player, the initial plan is NOT to add all these 500 features indiscriminately but to instead try to find useful features and incorporate these.

Scaling and power-transforms will not be used at this time since the plan is to use GBTs (gradient boosted trees) such as XGBoost where this is not needed. These transforms may be needed later for PCA and other techniques.


duplicates

 - both games.csv and ranking.csv contain several duplicated rows from Dec 2020 (covid season) that the pandas function *df.duplicated()* failed to detect in EDA. These will be filtered out using subsets instead of the entire dataframe.

 games.csv
 
 - delete preseason games (this will also take care of the null games from early 2003)
 - keep only games where GAME_STATUS_TEXT = 'Final' (for better utility in the future)
 - remove duplicated records 
 - flag postseason games 
 - drop 'GAME_STATUS_TEXT', 'TEAM_ID_home', 'TEAM_ID_away'

ranking.csv
 
 - drop preseason rankings (SEASON_ID begins with 1)
 - split HOME_RECORD into HOME_W, HOME_L, and HOME_W_PCT
 - split ROAD_RECORD into ROAD_W, ROAD_L, and ROAD_W_PCT
 - numericaly encode CONFERENCE (East or West)
 - remove duplicated records
 - drop 'SEASON_ID', 'LEAGUE_ID', 'RETURNTOPLAY', 'TEAM', 'HOME_RECORD', 'ROAD_RECORD'

 game_details.csv
 
 - fix mixed formats in MIN and convert to float
 - fix negatives in MIN
 - if MIN is null, edit START_POSITION to 'NP' (not played)
 - any START_POSITION remaining null, convert to NS (not start, but still played)
 - drop TEAM_ABBREVIATION, TEAM_CITY, PLAYER_NAME, NICKNAME, COMMENT
 
 Join games with ranking
 
  - LINK: games.GAME_DATE_EST, games.HOME_TEAM_ID, -> ranking.STANDINGSDATE, ranking.TEAM_ID 
  - ADD: CONFERENCE, G, W, L, W_PCT, HOME_W, HOME_L, HOME_W_PCT, ROAD_W, ROAD_L, ROAD_W_PCT
  - repeat with AWAY_TEAM_ID instead of HOME_TEAM_ID
  
 Add TARGET
 
  - Sort games by HOME_TEAM_ID and GAME_ID
  - for each SEASON and HOME_TEAM_ID, shift HOME_TEAM_WINS down to TARGET for previous game
  - remove games with null TARGETs (last game played each season by each team will have no null TARGET)
  

 

import pandas as pd
import numpy as np

pd.set_option('display.max_columns', 500)

# For Visualization
import matplotlib.pyplot as plt
import seaborn as sns

from pathlib import Path  #for Windows/Linux compatibility
DATAPATH = Path(r'data')


## games.csv

games = pd.read_csv(DATAPATH / "games.csv")
games.head()

**Clean Data**

#remove preseason games (GAME_ID begins with a 1)
games = games[games['GAME_ID'] > 20000000]

#flag postseason games (GAME_ID begins with >2)
games['PLAYOFF'] = (games['GAME_ID'] >= 30000000).astype('int8')

#remove duplicates (each GAME_ID should be unique)
games = games[~games.duplicated(subset=['GAME_ID'])]

#drop unnecessary fields
drop_fields = ['GAME_STATUS_TEXT', 'TEAM_ID_home', 'TEAM_ID_away']
games = games.drop(drop_fields,axis=1)

games
    

## ranking.csv

ranking = pd.read_csv(DATAPATH / "ranking.csv")
ranking.head()

**Clean Data**

#remove preseason rankings (SEASON_ID begins with 1)
ranking = ranking[ranking['SEASON_ID'] > 20000]

#convert home record and road record to numeric
ranking['HOME_W'] = ranking['HOME_RECORD'].apply(lambda x: x.split('-')[0]).astype('int')
ranking['HOME_L'] = ranking['HOME_RECORD'].apply(lambda x: x.split('-')[1]).astype('int')
ranking['HOME_W_PCT'] = ranking['HOME_W'] / ( ranking['HOME_W'] + ranking['HOME_L'] )

ranking['ROAD_W'] = ranking['ROAD_RECORD'].apply(lambda x: x.split('-')[0]).astype('int')
ranking['ROAD_L'] = ranking['ROAD_RECORD'].apply(lambda x: x.split('-')[1]).astype('int')
ranking['ROAD_W_PCT'] = ranking['ROAD_W'] / ( ranking['ROAD_W'] + ranking['ROAD_L'] )

#encode CONFERENCE as an integer (just using pandas - not importing sklearn for just one feature)
ranking['CONFERENCE'] = ranking['CONFERENCE'].apply(lambda x: 0 if x=='East' else 1 ).astype('int') 

#remove duplicates (there should only be one TEAM_ID per STANDINGSDATE)
ranking = ranking[~ranking.duplicated(subset=['TEAM_ID','STANDINGSDATE'])]

#drop unnecessary fields
drop_fields = ['SEASON_ID', 'LEAGUE_ID', 'RETURNTOPLAY', 'TEAM', 'HOME_RECORD', 'ROAD_RECORD']
ranking = ranking.drop(drop_fields,axis=1)

ranking


## game_details.csv

details = pd.read_csv(DATAPATH / "games_details.csv")
details.head()

**Clean Data**

# convert MIN:SEC to float
df = details.loc[details['MIN'].str.contains(':',na=False)]
df['MIN_whole'] = df['MIN'].apply(lambda x: x.split(':')[0]).astype("int8")
df['MIN_seconds'] = df['MIN'].apply(lambda x: x.split(':')[1]).astype("int8")
df['MIN'] = df['MIN_whole'] + (df['MIN_seconds'] / 60)

details['MIN'].loc[details['MIN'].str.contains(':',na=False)] = df['MIN']
details['MIN'] = details['MIN'].astype("float16")

# convert negatives to positive
details['MIN'].loc[details['MIN'] < 0] = -(details['MIN'])

#update START_POSITION if did not play (MIN = NaN)
details['START_POSITION'].loc[details['MIN'].isna()] = 'NP'

#update START_POSITION if null
details['START_POSITION'] = details['START_POSITION'].fillna('NS')

#drop unnecessary fields
drop_fields = ['COMMENT', 'TEAM_ABBREVIATION', 'TEAM_CITY', 'PLAYER_NAME', 'NICKNAME'] 
details = details.drop(drop_fields,axis=1)

details

## Add correct TARGET to dataframe

games['TARGET'] = games['HOME_TEAM_WINS']

games

**Save Train Data**

games.to_csv(DATAPATH / "transformed.csv",index=False)

**Sweetviz data visualization**

def run_sweetviz_report(df, TARGET):
    
    import sweetviz as sv
    from datetime import datetime
    
    report_label = datetime.today().strftime('%Y-%m-%d_%H_%M')
    
    my_report = sv.analyze(df,target_feat=TARGET)
    my_report.show_html(filepath='SWEETVIZ_' + report_label + '.html')
    
    return

df = pd.read_csv(DATAPATH / "transformed.csv")
run_sweetviz_report(df,'TARGET')

**Correlations with TARGET**

df = games.drop(columns=['GAME_ID', 'TARGET'])

x = df.corrwith(games['TARGET'], method = 'spearman').sort_values(ascending=False)

print(x)
