==> notebooks/06_feature_selection.ipynb <==
## Feature Selection

import pandas as pd
import numpy as np

pd.set_option('display.max_columns', 500)

from sklearn.feature_selection import SelectKBest, chi2


import matplotlib.pyplot as plt
import seaborn as sns

from pathlib import Path  #for Windows/Linux compatibility
DATAPATH = Path(r'data')


train = pd.read_csv(DATAPATH / "train_features.csv")
test = pd.read_csv(DATAPATH / "test_features.csv")


train.to_csv(DATAPATH / "train_selected.csv",index=False)
test.to_csv(DATAPATH / "test_selected.csv",index=False)
